chore: move getLambdaFunctionScanOutput to utils folder
