+++ src/utils/getLambdaFunctionScanOutput.spec.ts
@@ -3,14 +3,14 @@ import { rm } from "node:fs/promises";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
 import { getLambdaFunctionScanOutput } from "./getLambdaFunctionScanOutput.ts";
-import { downloadFile } from "./utils/downloadFile.ts";
-import { getLambdaFunctionContents } from "./utils/getLambdaFunctionContents.ts";
-import { hasSdkV2InBundle } from "./utils/hasSdkV2InBundle.ts";
+import { downloadFile } from "./downloadFile.ts";
+import { getLambdaFunctionContents } from "./getLambdaFunctionContents.ts";
+import { hasSdkV2InBundle } from "./hasSdkV2InBundle.ts";
 
 vi.mock("node:fs/promises");
-vi.mock("./utils/downloadFile.ts");
-vi.mock("./utils/getLambdaFunctionContents.ts");
-vi.mock("./utils/hasSdkV2InBundle.ts");
+vi.mock("./downloadFile.ts");
+vi.mock("./getLambdaFunctionContents.ts");
+vi.mock("./hasSdkV2InBundle.ts");
 
 describe("getLambdaFunctionScanOutput", () => {
   const mockClient = { getFunction: vi.fn() } as unknown as Lambda;
